Add support for assumed-role.
Here's an example STS assumed-role I created when attaching a role to an
EC2:

arn:aws:sts::251399821276:assumed-role/derp/i-075750a3a2378c42e

It looks them same when assuming a role in another account.

The last part of that slug is both user-controlled, and mostly random
if you use the default AWS SDK. As a result, we will transform the
`sts` slug for `assumed-role`, in order to not have that last bit.

This means the example above turns into

arn:aws:sts::251399821276:assumed-role/derp

This is slightly confusing, but hopefully something we can overcome with
documentation. User ARNs will continue through unchanged.