Make Quickstart Clearer
I wasted a few hours because I thought `/etc/ssh/hallow_cas` was a directory that was supposed to contain a filed called `hallow-ca.pub` with our pubkey. The result was ssh auth was failing in a very non-transparent way.

After studying the sshd error messages and lining them up with sshd C source code, I realized `TrustedUserCAKeys` is supposed to be a regular file. Figured I would pay it forward so the next person does not have to go through this experience :)