Correcting docs, which give wrong principal
The output from the AWS query to pull principal is wrong (at least in my situation, with assumed roles). The hallow-cli ssh-add doesn't include the account username in the principal, so principal lookups always fail. Since the certificate already has the principal written to it, I added a quick one-liner instead of the aws query.